backend/requirements.txt
asgiref==3.11.1
autobahn==24.4.2
Automat==25.4.16
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
isodate==0.7.2
lxml==6.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
postgrest==2.28.0
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
puremagic==1.30
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
PyMuPDF==1.27.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
pypdf==6.7.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
rdflib==7.6.0
realtime==2.28.0
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
storage3==2.28.0
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
whitenoise==6.9.0
zope.interface==8.2
